Add files via upload
You will notice within some of the columns, such as the column "Biological measure," there are some "Other" answers (eg, see row 15). The way DistillerSR works is we have a giant multiple choice sheet that we use to answer each of the extraction questions. When an answer is not an option, we manually write it in under "Other." 
At the end of data extraction, I will group these more "rare" answers into categories. For the purposes of creating the Sankey in the proposal, I just deleted the "Other" terms. This is the same with the "Outcome_comment" and "Domain_comment" columns.