Latest article information
This csv includes all extracted articles. We might need to delete a couple if we find problems but it is pretty much the final dataset. I've changed the names of some of the variables for consistency (sorry!). The "specific" option for Outcome and Behavior has a few more categories than the parent variable. I tried to use this dataset in my branch but I broke the app with it.